feat: 카드 삭제 기능 구현

+++ src/components/common/modal/modal-task/comment/index.tsx
@@ -40,7 +40,9 @@ const Comment = ({
   }
 
   const handleDelete = () => {
-    onDelete(commentId)
+    if (onDelete) {
+      onDelete(commentId)
+    }
   }
 
   return (
